+++ packages/widgetbook_generator/lib/code_generators/instances/addons/addons.dart
@@ -1,2 +1,4 @@
+export './addon_instance.dart';
+export './localization_addon/localization_addon.dart';
 export './material_theme_addon/material_theme_addon.dart';
 export './text_scale_addon/text_scale_addon.dart';