+++ packages/widgetbook_generator/lib/code_generators/instances/addons/localization_addon/localization_addon.dart
@@ -1,0 +1,2 @@
+export 'localization_addon_instance.dart';
+export 'localization_setting_instance.dart';